third_party/WebKit/LayoutTests/fast/dom/shadow/css-focus-pseudo-match-shadow-host-expected.txt

CONSOLE WARNING: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated. See https://www.chromestatus.com/features/4668884095336448 for more details.
PASS successfullyParsed is true

TEST COMPLETE
testing without tabstop for crbug/479050
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
testing single shadow DOM tree
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
testing shadow tree of trees
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(0, 0, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
testing distributed nodes
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
testing nested distribution
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
#input1, #input2 are (re)distributed in the same treescope as #input3, but :focus on shadow host only matches when a focused element is under its shadow tree.
Case 1: top and inner shadow do not delegate focus.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
Case 2: top shadow delegates, but inner shadow does not.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
Case 3: top shadow does not delegate, but inner shadow does.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(0, 0, 255)"
Case 4: both shadow hosts delagate focus.
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(255, 255, 255)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(255, 255, 0)"
PASS backgroundColorOf('shadow-host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('shadow-host/inner-shadow-host') is "rgb(0, 0, 255)"
